Version v0.6.18

+++ client/src/app/game/modals/unmet-condition-modal.tsx
@@ -1,31 +1,28 @@
 import Modal from 'components/modal'
-import {useDispatch} from 'react-redux'
-import css from './confirm-modal.module.css'
-
-import {removeEffect} from 'logic/game/game-actions'
+import css from './game-modals.module.scss'
 import Button from 'components/button'
+import {useDispatch} from 'react-redux'
+import {removeEffect} from 'logic/game/game-actions'
 
 type Props = {
 	closeModal: () => void
+	info: {removeSuAfter: boolean}
 }
-function UnmetCondition({closeModal}: Props) {
+function UnmetCondition({closeModal, info}: Props) {
 	const dispatch = useDispatch()
-
 	const handleOk = () => {
-		dispatch(removeEffect())
 		closeModal()
+		if (info?.removeSuAfter) {
+			dispatch(removeEffect())
+		}
 	}
 
 	return (
-		<Modal title="Unmet Condition">
+		<Modal title="Unmet Condition" closeModal={closeModal}>
 			<div className={css.confirmModal}>
-				<div className={css.description}>
-					You can't use this effect at the moment.
-				</div>
+				<div className={css.description}>You can't play this card at the moment.</div>
 				<div className={css.options}>
-					<Button variant="primary" size="small" onClick={handleOk}>
-						Ok
-					</Button>
+					<Button onClick={handleOk}>Okay</Button>
 				</div>
 			</div>
 		</Modal>
